Safari-609.1.11.1/LayoutTests/fast/text/line-height-minimumFontSize-text-small-font-size-expected.txt

PASS document.getElementById('target').offsetHeight < 100 is true
PASS successfullyParsed is true

TEST COMPLETE
Hi
